Work around what seems like a Cython bug
