+++ mathics/core/element.py
@@ -6,6 +6,7 @@ Here we have the base class and related function for element inside an Expressio
 """
 
 from mathics.core.attributes import no_attributes
+from recordclass import RecordClass
 from typing import Any, Optional, Tuple
 
 
@@ -33,6 +34,41 @@ def fully_qualified_symbol_name(name) -> bool:
         and not name.endswith("`")
         and "``" not in name
     )
+
+
+# Note: Something in cythonization barfs if we put this in Expression and you try to call this
+# like ExpressionProperties(True, True, True). Cython reports:
+#   number of the arguments greater than the number of the items
+class ElementsProperties(RecordClass):
+    """Properties of Expression elements that are useful in evaluation.
+
+    In general, if you have some set of properties that you know should
+    be set a particular way, but don't know about the others, it is safe
+    to set the unknown properties to False.
+
+    When *all* of the properties are unknown, use a `None` value in
+    the Expression.properties field instead of creating an
+    ElementsProperties object with everything set False.
+    """
+
+    # True if none of the elements needs to be evaluated.
+    elements_fully_evaluated: bool = False
+
+    # is_flat: True if none of the elements is an Expression
+    # Some Mathics functions allow flattening of elements. Therefore
+    # it can be useful to know if the elements are already flat
+    is_flat: bool = False
+
+    # is_ordered: True if all of the elements are ordered. Of course this is true,
+    # if there are less than 2 elements. Ordered is an Attribute of a
+    # Mathics function.
+    #
+    # In rewrite_eval_apply() if a function is not marked as Ordered this attribute
+    # has no effect which means it doesn't matter how it is set. So
+    # when it doubt, it is always safe to set is_ordered to False since at worst
+    # it will cause an ordering operation on elements sometimes. On the other hand, setting
+    # this True elements are not sorted can cause evaluation differences.
+    is_ordered: bool = False
 
 
 class ImmutableValueMixin:
